requirements.txt
pip
setuptools
wheel

python-dotenv

flask

sqlalchemy

pytest